# dj-connect-back/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'DJ Connect API'
  description: 'API documentation for DJ Connect. All requests must include header Telegram-Init-Data: ${TWA.initData()}'
  contact:
    name: 'Contact Developer'
    url: 'https://t.me/beilec'
  license:
    name: 'Contact Lead'
    url: 'https://t.me/alievdenis1'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8082/api/v1'
    description: 'DJ Connect API Server'
  -
    url: 'https://dj-connect.xyz/api/v1'
paths:
...
  '/transactions/{transaction_id}/mark-paid':
    patch:
      tags:
        - Transaction
      summary: 'Mark a transaction as paid'
      description: 'Allows a DJ to mark a transaction as paid'
      operationId: markTransactionPaid
      parameters:
        -
          name: transaction_id
          in: path
          description: 'Transaction ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Transaction marked as paid'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Transaction marked as paid'
                type: object
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'Transaction not found'
      security:
        -
          telegramAuth: []
  '/transactions/{transaction_id}/cancel':
    patch:
      tags:
        - Transaction
      summary: 'Cancel a transaction'
      description: 'Allows a DJ to cancel a pending transaction'
      operationId: cancelTransaction
      parameters:
        -
          name: transaction_id
          in: path
          description: 'Transaction ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Transaction cancelled successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Transaction cancelled'
                type: object
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'Transaction not found'
      security:
        -
          telegramAuth: []
  '/profile/{telegram_id}':
    get:
      tags:
        - User
      summary: "Get a user's profile by Telegram ID"
      description: 'Returns the details of a specified user by their Telegram ID'
      operationId: getUserProfileByTelegramId
      parameters:
        -
          name: telegram_id
          in: path
          description: 'Telegram ID of the user'
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: 'John Doe'
                  email:
                    type: string
                    example: john.doe@example.com
                  telegram_id:
                    type: string
                    example: '123456789'
                type: object
        404:
          description: 'User not found'
        401:
          description: Unauthorized
      security:
        -
          telegramAuth: []
  /profile/me:
    get:
      tags:
        - User
      summary: "Get the current authenticated user's details"
      description: 'Returns the details of the authenticated user'
      operationId: getCurrentUser
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: 'John Doe'
                  email:
                    type: string
                    example: john.doe@example.com
                  telegram_id:
                    type: string
                    example: '123456789'
                type: object
        401:
          description: Unauthorized
      security:
        -
          telegramAuth: []
components:
...
  securitySchemes:
    telegramAuth:
      type: apiKey
      in: header
      name: Telegram-Init-Data
      description: 'Value of TWA.initData() from Telegram Web App.'
...
security:
  -
    telegramAuth: []